requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain and AI dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.3.5
langchain_chroma>=0.1.2
langchain_core==0.3.60
langgraph==0.4.5

# Utility dependencies
beautifulsoup4>=4.12.2
nest_asyncio>=1.5.8
tqdm>=4.66.1
python-dotenv>=1.0.1
typing_extensions==4.13.2
protobuf<=3.20.3

# Database dependencies
chromadb>=0.4.18
duckdb>=0.8.1
# pysqlite3-binary - removed due to Docker compatibility issues
# ChromaDB will use system sqlite3 in Docker

# Optional: Keep streamlit for legacy support but not required for FastAPI
streamlit==1.45.1

